pyproject.toml
[project]
name = "langgraph-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-core>=0.3.68",
    "langchain-groq>=0.3.5",
    "langchain-openai>=0.3.27",
    "langchain-tavily>=0.2.6",
    "langgraph>=0.5.1",
    "python-dotenv>=1.1.1",
]
